Reuse field name tracking in OrcSchemaVisitor
